Update /pbv to account for an integer
Added if-else statements to attempt to clean the search for an integer. Searching for a Pokemon is now searching for "isNaN", as the searching mechanism for searching for a Pokemon is included already. 
Searching for a PBV (must be a multiple of 5 and not a decimal) now sends a reply box that links to the Pokedex that accounts for the limit specified as well as those below the limit. 
POSSIBLE ISSUES: 
(line 669) var pbvInput = target; (Is this legal?)